fix warnings

+++ src/store/benchmark/async/async_transaction_bench_client.h
@@ -29,7 +29,7 @@ private:
     bool retryAborted;
     int32_t maxAttempts;
     AsyncTransaction *currTxn;
-    size_t currTxnAttempts;
+    int32_t currTxnAttempts;
     std::mt19937 gen;
 
 };
